design onBoarding screen view
